fix findbugs

+++ src/main/java/com/example/spring/dao/CustomerDAO.java
@@ -15,11 +15,11 @@ public class CustomerDAO {
     // Dummy database. Initialize with some dummy values.
     private static List customers;
 
-    static{
+    static {
         customers = new ArrayList();
-        customers.add(new Customer(101, "John", "Doe", "djohn@gmail.com", "121-232-3435"));
-        customers.add(new Customer(201, "Russ", "Smith", "sruss@gmail.com", "343-545-2345"));
-        customers.add(new Customer(301, "Kate", "Williams", "kwilliams@gmail.com", "876-237-2987"));
+        customers.add(new Customer((long) 101, "John", "Doe", "djohn@gmail.com", "121-232-3435"));
+        customers.add(new Customer((long) 201, "Russ", "Smith", "sruss@gmail.com", "343-545-2345"));
+        customers.add(new Customer((long) 301, "Kate", "Williams", "kwilliams@gmail.com", "876-237-2987"));
     }
 
     /**
@@ -90,8 +90,8 @@ public class CustomerDAO {
      * Update the customer object for given id in dummy database. If customer
      * not exists, returns null
      *
-     * @param id
-     * @param customer
+     * @param id id
+     * @param customer customer
      * @return customer object with id
      */
     public Customer update(Long id, Customer customer) {
